Add supplier management feature with contacts support
- Implement SupplierService with CRUD for suppliers and contacts
- Register ISupplierService and SupplierValidator in DI container
- Update SupplierModel: use double for costs, add email/phone
- Extend SupplierPageViewModel and EntityPageViewModel for contact commands
- Expand SupplierView UI: contact tab, data grid, toolbar, record count
- Add navigation for supplier management in main/resources menu
- Add localization for supplier/contact UI and errors (NL, EN, DE)
- Add SupplierValidator for name/duplicate checks
- Update DBNames with new supplier/contact fields
- Increment project version and clean up duplicate registrations
- No breaking changes to existing features

+++ Modelbouwer/Models/SupplierModel.cs
@@ -28,16 +28,16 @@ public partial class SupplierModel : ObservableObject
 	private string? _supplierUrl;
 
 	[ObservableProperty]
-	private decimal _supplierShippingCosts;
+	private double _supplierShippingCosts;
 
 	[ObservableProperty]
-	private decimal _supplierMinShippingCosts;
+	private double _supplierMinShippingCosts;
 
 	[ObservableProperty]
-	private decimal _supplierOrderCosts;
+	private double _supplierOrderCosts;
 
 	[ObservableProperty]
-	private decimal _supplierMinOrderCosts;
+	private double _supplierMinOrderCosts;
 
 	[ObservableProperty]
 	private string? _supplierMemo;
